Moved IBindable.cs under the Bindables folder
