+++ Assets/FastenUp/Runtime/Delegates/ValueChangedDelegate.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using FastenUp.Runtime.Base;
+using FastenUp.Runtime.Bindables;
 
 namespace FastenUp.Runtime.Delegates
 {
